xds: move LocalTCPListener to internal/testutils
This will be used from RLS tests.

RELEASE NOTES: none
